requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
astor==0.7.1
automium==0.2.6
automium-web==0.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.2
Flask-Classful==0.14.2
Flask-Dropzone==1.5.3
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
gast==0.2.2
get==2018.11.19
grpcio==1.19.0
h5py==2.9.0
idna==2.8
imageai==2.0.2
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
legacy==0.1.6
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.24.2
pathlib==1.0.1
pbr==5.1.3
Pillow==5.4.1
post==2018.11.20
protobuf==3.7.0
public==2018.11.20
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
scipy==1.2.1
six==1.12.0
SQLAlchemy==1.3.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.1
views==0.3
Werkzeug==0.14.1
wget==3.2
